+++ packages/headless/src/features/facets/automatic-facet-set/automatic-facet-set-slice.test.ts
@@ -1,17 +1,22 @@
+import {AutomaticFacetGeneratorOptions} from '../../../controllers/facets/automatic-facet-generator/headless-automatic-facet-generator-options';
 import {buildMockAutomaticFacetResponse} from '../../../test/mock-automatic-facet-response';
+import {buildMockAutomaticFacetSlice} from '../../../test/mock-automatic-facet-slice';
 import {buildMockFacetValue} from '../../../test/mock-facet-value';
 import {buildMockSearch} from '../../../test/mock-search';
 import {logSearchEvent} from '../../analytics/analytics-actions';
+import {deselectAllBreadcrumbs} from '../../breadcrumb/breadcrumb-actions';
+import {restoreSearchParameters} from '../../search-parameters/search-parameter-actions';
 import {executeSearch} from '../../search/search-actions';
 import {FacetValueState} from '../facet-api/value';
 import {
   deselectAllAutomaticFacetValues,
-  setDesiredCount,
+  setOptions,
   toggleSelectAutomaticFacetValue,
 } from './automatic-facet-set-actions';
 import {automaticFacetSetReducer} from './automatic-facet-set-slice';
 import {
   AutomaticFacetSetState,
+  AutomaticFacetSlice,
   getAutomaticFacetSetInitialState,
 } from './automatic-facet-set-state';
 import {AutomaticFacetResponse} from './interfaces/response';
@@ -44,24 +49,38 @@ describe('automatic-facet-set slice', () => {
       const facets: AutomaticFacetResponse[] = [
         buildMockAutomaticFacetResponse(),
       ];
-      const facetsRecord: Record<string, AutomaticFacetResponse> = {
-        [facets[0].field]: facets[0],
+      const facetsRecord: Record<string, AutomaticFacetSlice> = {
+        [facets[0].field]: {response: facets[0]},
       };
       const action = buildExecuteSearchAction(facets);
 
       const finalState = automaticFacetSetReducer(state, action);
 
-      expect(finalState.facets).toEqual(facetsRecord);
+      expect(finalState.set).toEqual(facetsRecord);
     });
   });
 
-  it('sets state.desiredCount when #setDesiredCount', () => {
-    const desiredCount = 5;
-    const action = setDesiredCount(desiredCount);
+  it('sets state.desiredCount when #setOptions', () => {
+    const options: AutomaticFacetGeneratorOptions = {desiredCount: 8};
+
+    const action = setOptions(options);
 
     const finalState = automaticFacetSetReducer(state, action);
 
-    expect(finalState.desiredCount).toEqual(desiredCount);
+    expect(finalState.desiredCount).toEqual(options.desiredCount);
+  });
+
+  it('sets state.numberOfValues when #setOptions', () => {
+    const options: AutomaticFacetGeneratorOptions = {
+      desiredCount: 8,
+      numberOfValues: 6,
+    };
+
+    const action = setOptions(options);
+
+    const finalState = automaticFacetSetReducer(state, action);
+
+    expect(finalState.numberOfValues).toEqual(options.numberOfValues);
   });
 
   describe('#toggleSelectAutomaticFacetValue', () => {
@@ -73,14 +92,16 @@ describe('automatic-facet-set slice', () => {
 
       const finalState = automaticFacetSetReducer(state, action);
 
-      expect(finalState.facets).toEqual(state.facets);
+      expect(finalState.set).toEqual(state.set);
     });
 
     it('does nothing if it cannot find the value', () => {
-      const facet = buildMockAutomaticFacetResponse({
+      const response = buildMockAutomaticFacetResponse({
         values: [buildMockFacetValue({value: 'valuePresent'})],
       });
-      state.facets = {[facet.field]: facet};
+      const slice = buildMockAutomaticFacetSlice({response});
+
+      state.set = {[response.field]: slice};
       const action = toggleSelectAutomaticFacetValue({
         field: 'fieldPresent',
         selection: buildMockFacetValue({value: 'valueNotPresent'}),
@@ -88,7 +109,7 @@ describe('automatic-facet-set slice', () => {
 
       const finalState = automaticFacetSetReducer(state, action);
 
-      expect(finalState.facets).toEqual(state.facets);
+      expect(finalState.set).toEqual(state.set);
     });
 
     describe('toggles the value if', () => {
@@ -101,11 +122,13 @@ describe('automatic-facet-set slice', () => {
         });
       }
       function addFacetToState(valueState: FacetValueState) {
-        const facet = buildMockAutomaticFacetResponse({
+        const response = buildMockAutomaticFacetResponse({
           field,
           values: [buildMockFacetValue({value, state: valueState})],
         });
-        state.facets[field] = facet;
+        const slice = buildMockAutomaticFacetSlice({response});
+
+        state.set[field] = slice;
       }
 
       it('is "selected"', () => {
@@ -113,7 +136,7 @@ describe('automatic-facet-set slice', () => {
         const action = buildToggleSelectAutomaticFacetValueAction();
 
         const finalState = automaticFacetSetReducer(state, action);
-        const targetValue = finalState.facets[field].values.find(
+        const targetValue = finalState.set[field].response.values.find(
           (req) => req.value === value
         );
 
@@ -125,7 +148,7 @@ describe('automatic-facet-set slice', () => {
         const action = buildToggleSelectAutomaticFacetValueAction();
 
         const finalState = automaticFacetSetReducer(state, action);
-        const targetValue = finalState.facets[field].values.find(
+        const targetValue = finalState.set[field].response.values.find(
           (req) => req.value === value
         );
 
@@ -140,23 +163,76 @@ describe('automatic-facet-set slice', () => {
 
       const finalState = automaticFacetSetReducer(state, action);
 
-      expect(finalState.facets).toEqual(state.facets);
+      expect(finalState.set).toEqual(state.set);
     });
 
     it('puts all values to "idle"', () => {
       const field = 'field';
-      const facet = buildMockAutomaticFacetResponse({
+      const response = buildMockAutomaticFacetResponse({
         field,
         values: [
           buildMockFacetValue({value: 'value1', state: 'selected'}),
           buildMockFacetValue({value: 'value2', state: 'selected'}),
         ],
       });
-      state.facets[field] = facet;
+      const slice = buildMockAutomaticFacetSlice({response});
+
+      state.set[field] = slice;
       const action = deselectAllAutomaticFacetValues(field);
 
       const finalState = automaticFacetSetReducer(state, action);
-      const targetValues = finalState.facets[field].values;
+      const targetValues = finalState.set[field].response.values;
+
+      expect(targetValues.every((value) => value.state === 'idle')).toEqual(
+        true
+      );
+    });
+  });
+
+  describe('#restoreSearchParameters', () => {
+    it('sets #values to the selected values in the payload when a facet is found in the #af payload', () => {
+      const field = 'field';
+      const value = 'a';
+      const af = {[field]: [value]};
+      const action = restoreSearchParameters({af});
+
+      const finalState = automaticFacetSetReducer(state, action);
+      const selectedValue = buildMockFacetValue({value, state: 'selected'});
+
+      expect(finalState.set[field].response.values).toEqual([selectedValue]);
+    });
+  });
+
+  describe('#deselectAllBreadcrumbs', () => {
+    it('sets all responses #values to "idle"', () => {
+      const field1 = 'field1';
+      const field2 = 'field2';
+      const response1 = buildMockAutomaticFacetResponse({
+        field: field1,
+        values: [
+          buildMockFacetValue({value: 'value1', state: 'selected'}),
+          buildMockFacetValue({value: 'value2', state: 'selected'}),
+        ],
+      });
+      const response2 = buildMockAutomaticFacetResponse({
+        field: field2,
+        values: [
+          buildMockFacetValue({value: 'value1', state: 'selected'}),
+          buildMockFacetValue({value: 'value2', state: 'selected'}),
+        ],
+      });
+      const slice1 = buildMockAutomaticFacetSlice({response: response1});
+      const slice2 = buildMockAutomaticFacetSlice({response: response2});
+
+      state.set[field1] = slice1;
+      state.set[field2] = slice2;
+      const action = deselectAllBreadcrumbs();
+
+      const finalState = automaticFacetSetReducer(state, action);
+      const targetValues = [
+        ...finalState.set[field1].response.values,
+        ...finalState.set[field2].response.values,
+      ];
 
       expect(targetValues.every((value) => value.state === 'idle')).toEqual(
         true